Fixed buggy chart command. Organized commands to seperate cogs for increased maintainability
