+++ .pre-commit-config.yaml
@@ -11,7 +11,7 @@ repos:
     hooks:
       - id: mypy
   - repo: https://github.com/psf/black
-    rev: 22.1.0
+    rev: 22.3.0
     hooks:
       - id: black
   - repo: https://github.com/MarcoGorelli/absolufy-imports
