+++ setup.cfg
@@ -1,6 +1,6 @@
 [flake8]
 show-source = true
-ignore = WPS305, DAR201, DAR101, C812, WPS608, D100, WPS347, I005, I001, WPS237, WPS316, WPS210, WPS605, WPS602, WPS323, WPS306, D101, WPS336, WPS229, WPS339, WPS432, F722, WPS320, WPS529, WPS114, WPS201, P101, WPS221, WPS231, WPS221, C901, WPS226, WPS232, WPS100, WPS407, WPS515
+ignore = WPS305, DAR201, DAR101, C812, WPS608, D100, WPS347, I005, I001, WPS237, WPS316, WPS210, WPS605, WPS602, WPS323, WPS306, D101, WPS336, WPS229, WPS339, WPS432, F722, WPS320, WPS529, WPS114, WPS201, P101, WPS221, WPS231, WPS221, C901, WPS226, WPS232, WPS100, WPS407, WPS515, WPS217
 max-line-length = 88
 max-complexity = 6
 inline-quotes = double
